app/requirements.txt
django==4.2.7
Markdown==3.5.1
djangorestframework==3.14.0
django-filter==2.4.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7

